EDU-801 Extract technical maintenance tab on admin page
