+++ src/components/admin/settings-tab.less
@@ -6,24 +6,6 @@
   &:first-of-type {
     margin-top: 0;
   }
-
-  &.SettingsTab-collapse--danger {
-    border-color: @edu-error-color;
-
-    & > .ant-collapse-item {
-      border-color: @edu-error-color;
-    }
-  }
-}
-
-.SettingsTab-collapseRow {
-  display: flex;
-  justify-content: space-between;
-  margin-bottom: 20px;
-
-  &:last-of-type {
-    margin-bottom: 0;
-  }
 }
 
 .SettingsTab-collapsTabPane {
