+++ src/components/admin/settings-tab.yml
@@ -13,30 +13,12 @@ footerLinksHeader:
 consentHeader:
   en: Consent text
   de: Zustimmungstext
-documentRegenerationButton:
-  en: Regenerate documents
-  de: Dokumente neu erstellen
-documentValidationButton:
-  en: Validate documents
-  de: Dokumente validieren
-cdnResourcesConsolidationButton:
-  en: Consolidate CDN resources
-  de: CDN-Ressourcen konsolidieren
-cdnUploadDirectoryCreationButton:
-  en: Create CDN upload directories
-  de: CDN-Upload-Verzeichnisse erstellen
 templateDocumentHeader:
   en: Template for new documents
   de: Vorlage für neue Dokumente
 licenseHeader:
   en: License
   de: Lizenz
-technicalMaintenanceTasksHeader:
-  en: Technical maintenance tasks
-  de: Technische Instandhaltungsaufgaben
-lastExecution:
-  en: Last execution
-  de: Letzte Ausführung
 pluginsHelpTextsHeader:
   en: Help texts for plugins
   de: Hilfetexte für Plugins
