+++ src/components/admin/technical-maintenance-tab.less
@@ -1,0 +1,11 @@
+.TechnicalMaintenanceTab {}
+
+.TechnicalMaintenanceTab-row {
+  display: flex;
+  justify-content: space-between;
+  margin-bottom: 20px;
+
+  &:last-of-type {
+    margin-bottom: 0;
+  }
+}
