+++ src/components/pages/admin.yml
@@ -4,3 +4,6 @@ settingsTabTitle:
 storagePlansTabTitle:
   en: Storage plans
   de: Speicherpläne
+technicalMaintenanceTabTitle:
+  en: Technical maintenance
+  de: Technische Instandhaltung
