+++ src/resources/resources.json
@@ -50,14 +50,8 @@
       "termsPageHeader": "Terms of usage page",
       "footerLinksHeader": "Footer links",
       "consentHeader": "Consent text",
-      "documentRegenerationButton": "Regenerate documents",
-      "documentValidationButton": "Validate documents",
-      "cdnResourcesConsolidationButton": "Consolidate CDN resources",
-      "cdnUploadDirectoryCreationButton": "Create CDN upload directories",
       "templateDocumentHeader": "Template for new documents",
       "licenseHeader": "License",
-      "technicalMaintenanceTasksHeader": "Technical maintenance tasks",
-      "lastExecution": "Last execution",
       "pluginsHelpTextsHeader": "Help texts for plugins"
     }
   },
@@ -70,14 +64,8 @@
       "termsPageHeader": "Nutzungsbedingungen-Seite",
       "footerLinksHeader": "Footer-Links",
       "consentHeader": "Zustimmungstext",
-      "documentRegenerationButton": "Dokumente neu erstellen",
-      "documentValidationButton": "Dokumente validieren",
-      "cdnResourcesConsolidationButton": "CDN-Ressourcen konsolidieren",
-      "cdnUploadDirectoryCreationButton": "CDN-Upload-Verzeichnisse erstellen",
       "templateDocumentHeader": "Vorlage für neue Dokumente",
       "licenseHeader": "Lizenz",
-      "technicalMaintenanceTasksHeader": "Technische Instandhaltungsaufgaben",
-      "lastExecution": "Letzte Ausführung",
       "pluginsHelpTextsHeader": "Hilfetexte für Plugins"
     }
   },
@@ -115,6 +103,28 @@
     "resources": {
       "assignedUserCount": "Benutzerkonten",
       "sortedByAssignedUserCount": "Nach Benutzerkonten sortiert"
+    }
+  },
+  {
+    "namespace": "technicalMaintenanceTab",
+    "language": "en",
+    "resources": {
+      "lastExecution": "Last execution",
+      "documentRegenerationButton": "Regenerate documents",
+      "documentValidationButton": "Validate documents",
+      "cdnResourcesConsolidationButton": "Consolidate CDN resources",
+      "cdnUploadDirectoryCreationButton": "Create CDN upload directories"
+    }
+  },
+  {
+    "namespace": "technicalMaintenanceTab",
+    "language": "de",
+    "resources": {
+      "lastExecution": "Letzte Ausführung",
+      "documentRegenerationButton": "Dokumente neu erstellen",
+      "documentValidationButton": "Dokumente validieren",
+      "cdnResourcesConsolidationButton": "CDN-Ressourcen konsolidieren",
+      "cdnUploadDirectoryCreationButton": "CDN-Upload-Verzeichnisse erstellen"
     }
   },
   {
@@ -766,7 +776,8 @@
     "language": "en",
     "resources": {
       "settingsTabTitle": "Settings",
-      "storagePlansTabTitle": "Storage plans"
+      "storagePlansTabTitle": "Storage plans",
+      "technicalMaintenanceTabTitle": "Technical maintenance"
     }
   },
   {
@@ -774,7 +785,8 @@
     "language": "de",
     "resources": {
       "settingsTabTitle": "Einstellungen",
-      "storagePlansTabTitle": "Speicherpläne"
+      "storagePlansTabTitle": "Speicherpläne",
+      "technicalMaintenanceTabTitle": "Technische Instandhaltung"
     }
   },
   {
