+++ src/styles/main.less
@@ -103,6 +103,7 @@
 @import '../components/media-player/chapter-selector.less';
 @import '../components/pages/complete-password-reset.less';
 @import '../components/media-player/media-player-track.less';
+@import '../components/admin/technical-maintenance-tab.less';
 @import '../components/media-player/media-volume-slider.less';
 @import '../components/resource-picker/storage-location.less';
 @import '../components/media-player/media-range-selector.less';
